+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 **Improvements/Fixes**
 - Fixed a bug where the Personal Compactor tooltip would get stuck on screen when using inventory search highlighting.
 - Fixed a bug where the calculator answer text would disappear if you moved the cursor from the end of the line to edit a number in the middle of the line.
+- Improved the Mutastion Recipe type.